requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
Django==4.0
autopep8==1.6.0
certifi==2021.10.8
configobj==5.0.6
django-cors-headers==3.10.0
djangorestframework==3.12.4
gunicorn==20.1.0
pycodestyle==2.8.0
python-dateutil==2.8.2
pytz==2021.3
sib-api-v3-sdk==7.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.7